# backend/app/api/alerts.py
"""
Alert API endpoints

Implements alert management according to OpenAPI specification.
"""

from fastapi import APIRouter, Query
from typing import Optional, List

router = APIRouter()


@router.get("")
async def get_alerts(
    limit: int = Query(50, le=500),
    offset: int = Query(0, ge=0),
    severity: Optional[List[str]] = Query(None),
    camera_id: Optional[int] = Query(None),
    start_time: Optional[str] = Query(None),
    end_time: Optional[str] = Query(None)
):
    """
    Get alerts

    Retrieve alert history with pagination and filtering.
    """
    # TODO: Implement alert retrieval from database
    return {
        "success": True,
        "data": [],
        "total": 0,
        "hasMore": False
    }


@router.get("/{alert_id}")
async def get_alert(alert_id: str):
    """Get alert by ID"""
    # TODO: Implement
    return {
        "success": True,
        "data": {
            "id": alert_id,
            "cameraId": 1,
            "type": "person_detected",
            "severity": "medium",
            "timestamp": "2024-01-12T10:00:00Z",
            "status": "new"
        }
    }


@router.post("/{alert_id}/acknowledge")
async def acknowledge_alert(alert_id: str):
    """Acknowledge alert"""
    # TODO: Implement
    return {
        "success": True,
        "data": {
            "id": alert_id,
            "status": "acknowledged"
        }
    }