# backend/app/api/search.py
"""
Search API endpoints
"""

from fastapi import APIRouter, Query
from typing import Optional, List

router = APIRouter()


@router.get("/videos")
async def search_videos(
    query: str = Query(...),
    camera_ids: Optional[List[int]] = Query(None),
    start_time: str = Query(...),
    end_time: str = Query(...),
    limit: int = Query(20)
):
    """
    Search recorded videos

    Search through recorded CCTV footage.
    """
    # TODO: Implement video search
    return {
        "success": True,
        "data": []
    }


@router.get("/events")
async def search_events(
    event_type: Optional[List[str]] = Query(None),
    camera_ids: Optional[List[int]] = Query(None),
    start_time: str = Query(...),
    end_time: str = Query(...)
):
    """
    Search events

    Search through detection events (motion, person, vehicle, etc.)
    """
    # TODO: Implement event search
    return {
        "success": True,
        "data": []
    }